requirements.txt
colorama
requests
dnspython
aiohttp
ipwhois
python-whois
